backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
celery==5.3.6
redis==5.0.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
stripe==8.6.0
google-generativeai>=0.8.3
pandas==2.2.1
python-docx==1.1.0
python-dotenv==1.0.1
xlsxwriter==3.2.0
openpyxl==3.1.2
